changing pid doubles to floats and testing size of protobug messages

+++ MotorDriver/lib/IRISMotorDriver/include/PIDHandler.h
@@ -19,11 +19,11 @@ class PIDHandler{ //rough outline
     public:
         PIDHandler();
 
-        double get_motor_value() const {return output_;}
-        //void set_encoder_value(double encoder_value);
-        void set_new_setpoint(double setpoint);
+        float get_motor_value() const {return output_;}
+        //void set_encoder_value(float encoder_value);
+        void set_new_setpoint(float setpoint);
 
-        bool update_pid(double input);
+        bool update_pid(float input);
         //bool get_inverted() {return motor_pid.GetDirection();}
         void setEnabled(bool enabled);
         bool getEnabled() const {return enabled;}
@@ -36,10 +36,10 @@ class PIDHandler{ //rough outline
         byte get_motorID() const {return motor_ID_;}
         byte get_encoderID() const {return encoder_ID_;}
         //for some reason the get methods for PID are not const. they are literaly const, but they are not marked as such 
-        double get_kd() {return motor_pid.GetKd();}
-        double get_ki() {return motor_pid.GetKi();}
-        double get_kp() {return motor_pid.GetKd();}
-        double get_setPoint() const {return setpoint_;}
+        float get_kd() {return motor_pid.GetKd();}
+        float get_ki() {return motor_pid.GetKi();}
+        float get_kp() {return motor_pid.GetKd();}
+        float get_setPoint() const {return setpoint_;}
 
         bool applyConfigUpdate(const PID_Config_Data& update);
         bool applySetPoint(const Set_PID_Setpoint& update_setpoint);
@@ -50,9 +50,9 @@ class PIDHandler{ //rough outline
     private:
         bool enabled;
 
-        double input_ = 0;
-        double output_ = 0; 
-        double setpoint_ = 0;
+        float input_ = 0;
+        float output_ = 0; 
+        float setpoint_ = 0;
 
         bool in_control;
         byte motor_ID_;
